Implement interface DataRecordsWriter.
